+++ packages/logic/src/hooks/api/index.ts
@@ -8,3 +8,4 @@ export { useCombinations } from './use-combinations';
 export { useDevelopers } from './use-developers';
 export { useFilmdevImport } from './use-filmdev-import';
 export { useFilms } from './use-films';
+export { useStats } from './use-stats';